+++ src/app/cart/cart.module.ts
@@ -1,5 +1,6 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
+import { FormsModule } from '@angular/forms';
 
 
 import { SharedModule } from '../shared/shared.module';
@@ -16,7 +17,8 @@ import { CartItemComponent } from './components/cart-item/cart-item.component';
   ],
   imports: [
     CommonModule,
-    SharedModule
+    SharedModule,
+    FormsModule
   ]
 })
 export class CartModule { }