+++ src/app/products/components/product-list/product-list.component.ts
@@ -1,4 +1,5 @@
 import { Component, OnInit } from '@angular/core';
+import { Observable } from 'rxjs';
 
 import { ProductModel } from '../../models/product.model';
 import { ProductsService } from '../../services/products.service';
@@ -11,7 +12,7 @@ import { CommunicateService } from 'src/app/cart/services/communicate.service';
 })
 export class ProductListComponent implements OnInit {
 
-  products: Array<ProductModel>;
+  products: Observable<Array<ProductModel>>;
 
   constructor(private productsService: ProductsService, private communicateService: CommunicateService) { }