+++ src/app/shared/shared.module.ts
@@ -3,18 +3,21 @@ import { CommonModule } from '@angular/common';
 
 import { BackgroundColorDirective } from './directives/background-color.directive';
 import { BorderDirective } from './directives/border.directive';
+import { OrderByPipe } from './pipes/order-by.pipe';
 
 @NgModule({
   declarations: [
     BackgroundColorDirective,
-    BorderDirective
+    BorderDirective,
+    OrderByPipe
   ],
   imports: [
     CommonModule
   ],
   exports: [
     BackgroundColorDirective,
-    BorderDirective
+    BorderDirective,
+    OrderByPipe
   ]
 })
 export class SharedModule { }